backend/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
pydantic==2.12.4
python-dotenv==1.2.1
alembic==1.17.2
python-multipart==0.0.20
mypy==1.18.2
flake8==7.3.0
black==25.11.0
isort==7.0.0
pytest==9.0.1
httpx==0.28.1
asyncpg==0.30.0
greenlet==3.2.4
trio==0.22.2